StartEditingDialog: Convert to ProgressDialog for mobile support
On mobile, this dialog has a title bar, and the back/forward buttons go
in the title bar (header). This is easily accomplished by converting it
to a ProcessDialog. On desktop, it still looks the same as it did before:
no header, and the back/forward buttons are at the bottom. This took
a bit of hackery.

On desktop, we:
- Make the close and back actions label-only instead of icon-only
- Hide the title bar using CSS, and move the body up accordingly
- Remove the outline from the footer
- Move all actions that would normally go in the header to the footer
  (which was already being done, just slightly differently)
- Remove the actionWidget class from these actions, to avoid weird and
  super specific OOUI styles that would make them look weird and ugly

Bug: T236854
Change-Id: Ibb4582580d8de64da62855f82d2b632cc1545a23