StartEditingDialog: Make sure vector-view-edit message is exported
This is a bit of a hack, and a better long-term solution might be to get
the caption of the edit tab and inject it as a parameter, and/or to have
a separate message for mobile where the edit button is a pencil icon
without a label. But this will fix the most urgent issue for now.

Bug: T237710
Change-Id: I92f7b02af9876a3975a6bd0694d6e8b992c11d50