Newcomer tasks: Skip fetching the extract if there's no API for it
Convenient for local debugging, and probably makes life easier if
someone wants to use the extension on a non-Wikimedia wiki.
(Note: setting $wgGERestbaseUrl to false, which is the default,
will result in trying to use RESTBase at its default location;
setting it to null will result in skipping the API call.)

Ideally the extracts API caller should be an injectable service,
but we don't really have conventions for doing that in Javascript.

Change-Id: I487b4fe0b3f44f3c5843f6c3a7ae96c93464c1d2