Use AQS for fetching pageview data
The query+pageviews API fails to return data for more than a
handful of pages. To some extent this is probably a bug in
the PageViewInfo extension, but returning data for 250 pages
would not be feasible in any case.

As a quick hack, use direct on demand AQS calls instead, reusing
the logic we already have for PCS summary calls. Use configuration
from PageViewInfo for determining the AQS URL. While we are there,
use that configuration for EditInfoService as well.

In the longer term, we'll probably want to use the action API
and fetch the information in small batches (at that point we can
also fix the use of query+pageimages which has a limit of 50 titles
per request), but PageViewInfo needs to be fixed first, plus that's
a less trivial change in the client-side logic.

Bug: T238178
Change-Id: I94cc92c1ba7310e2b7a9c6696901487dea56de59