Suggested edits: style filter buttons toolbar-like on mobile
On desktop, the button is expected to behave like module
content (honor module wrapper padding etc), on mobile it is
expected to behave like part of the block frame (snap to the
edges). There are a number of ways to do this, all bad:
* split mobile and desktop HTML
* use JS manipulation to relocate the container on mobile
* use negative margins (which will end up as some calc()-based
  mix of absolute and relative lengths)
* use absolute positioning
This patch takes the first approach (with a hacky method for
outputting content outside the normal body area, which seemed
still less disruptive than overriding the base render methods).

Bug: T238460
Change-Id: I6a268016c554f280e7ddd1d6ce58669f88fd35bb