Allow auto-enable for help panel to be tied to homepage
If $wgGEHelpPanelNewAccountEnableWithHomepage is true, the enable
percentage for the help panel will be ignored, and instead the help
panel will be enabled for new accounts that get the homepage. This will
allow us to align the treatment groups for the help panel and the
homepage, and always enable either both or neither.

Bug: T232396
Change-Id: Ibcc979e50a7718d9912dd3555ad7946e0fff48ef