SuggestedEdits: Allow pre-initiating a percentage of new accounts
For the specified percentage of newly created accounts, the suggested
edits module will be pre-initiated: we'll pretend that the user has
already clicked the button in the startediting module, gone through the
dialog and enabled the feature. When the user visits their homepage,
they won't see the startediting module, but instead see the suggested
edits module right away.

The percentage in the config variable is not the percentage of overall
new accounts, but a percentage of the accouns that win the "homepage on
new account" lottery. In other words, if
$wgGEHomepageNewAccountEnablePercentage = 50 and
$wgGEHomepageSuggestedEditsNewAccountInitiatedPercentage = 30, then 15%
of all new accounts will get suggested edits pre-initiated (30% of the
50% that get the homepage at all).

Bug: T238888
Change-Id: I3a0cf41669ebd18cf1b390f37fc61963a9f3825a