Initialize suggested edits module on the client side
Instead of reloading the page when the SuggestedEdits module is
enabled, rearrange/update the other modules manually from the JS
code (this results in some duplication of logic, but it's not
too bad), then load the module using the mechanism established
in I26bb166320.

Update promise handling slightly in Homepage.SuggestedEdits and
export the promise for the initial load, and use that in
StartEditingDialog to keep the dialog open until everything is
set up, to hide the jarring rapid sequence of visual changes
from the user. This is currently broken on mobile, as the dialog
closes on navigation (presumably this is some MobileFrontend
behavior); figuring that out is left to another patch.

Bug: T238179
Change-Id: I29a058b853d893799c9557632a4b52d33d00e996