Implement local configuration loader
Rename the RemotePageConfigurationLoader to PageConfigurationLoader, and perform
a check as to whether the config title contains an interwiki link. If so,
attempt to load it remotely. Otherwise load it locally via the DB rather than as
an HTTP request.

Existing configuration without interwiki links will switch to local page loading
when this patch is deployed. The existing GENewcomerTasksRemoteConfigTitle
variable is marked as deprecated in extension.json.

Bug: T235717
Change-Id: I5b68d33aaa64773c27ffa7f9d8e6e5539de1c6f0