Add newcomer task tag when user makes edit
Implement a server-side storage based tracker for keep track of whether a
successful edit originated from the user clicking on a task in the Suggested
Edits module.

The URL for a card is rewritten to go to
Special:Homepage/newcomertasks/{pageId}. SpecialHomepage.php is then responsible
for redirecting the user to the article page and preserving geclickid which is
used for joining with data from EditAttemptStep schema. Before redirecting,
it will use a new Tracker service for setting the page ID in cache for that
user's ID.

Later, in the RecentChanges save hook, the Tracker service is used again to
see if the RC object's article ID is in the store for that user, and if so, we
can mark the edit with a "newcomer task" tag.

The TTL is set to 1 week after looking at the discussion on T236885, but we
could adjust it down. The page ID is deliberately not removed from cache in
the RC changes hook in case the user makes subsequent edits; the cache TTL
will be solely responsible for removing the ID.

The previous click ID approach is left in place since that gives us data about
failed attempts to edit.

Bug: T236885
Change-Id: I3cf5abce18712d3ddd3eba10191abee70827c794