Suggested edits mobile (item E): Adjust styling for previous/next
Set left/right padding to 16px and icon size to 20px.

While the buttons have padding of 16px, the left button looks off center because
of the placement of the arrow icon within the button. That can either be dealt
with in another revision to this patch or in a follow-up patch.

Bug: T238280
Change-Id: I63bac751bb906c4188003aba335a11316d4076c4