Suggested edits: Don't fail so hard if pcs/aqs data are undefined
When the query to RESTBase and/or pageviews API results in a 404, pcsData and/or
aqsData are undefined. This causes an error to occur, which the end user
experiences as clicking/tapping a previous/next arrow but not seeing the UI
update, then clicking/tapping again and seeing a jump to the card before/after
next.

Change-Id: I1d62caff2200e28ee5ddf6a02e53c15b8304dc3c