Make PageViewInfo a soft dependency
PageViewInfo is supposed to be treated as a soft dependency, but the
homepage and Special:Impact break if it is disabled due to how services are
passed to SpecialHomepage. This commit uses a closure to instantiate
SpecialHomepage with the correct order of service arguments (thanks Daniel K for
the tip).

Bug: T240444
Change-Id: Id7ab14f4add1935c0efca3532e3dcf06ff15046a
(cherry picked from commit 35c11ec8a72e084d360d5491dc4549fcfffd04cb)