Suggested Edits: Remove close button from widget, close on hover out
Partially fixes T238165 by removing the 'head' on the popup widget (which comes
with a close button) and adding the label + description in the content. Then,
detect if a user hovered in, and if a hover out occurs on the popup, close the
popup.

Tapping the "?" button will also still close the popup.

Still needed is for the popup to disappear if the user moves the mouse away from
the direction of the popup area after it has first appeared, like in Page
Previews.

Bug: T238165
Change-Id: Ia22db75b6667228c58d57488094d85e7ed7bf883