Allow modules to export arbitrary module data, not just overlays
Replaces HomepageModule::getDataForOverlay() with getJsData() which
is handled similarly but is not specific to mobile summary/overlay
mode, and can also be used for things like dynamic module activation.

Use it to export module HTML for SuggestedEdits module
(when enabled but not activated).

Simplify the SpecialHomepage logic slightly by not returning the
rendered module list from the render* methods (it wasn't really
used for anything, and it contained modules with canRender() false).

Bug: T238179
Change-Id: I26bb166320b7f7c782107246b6ba412353c1c064