Simplify TrackerFactory / SpecialHomepage dependency injection
We don't need to check if suggested edits is enabled for the user to
inject the tracker factory service. We can check if its enabled before handling
a possible newcomer task redirection.

We also don't need to check if the user is logged in (already happens at top of
SpecialHomepage.php).

Bug: T240997
Change-Id: Icfae63bec4084bd4308512d803bdab406d4a9c18