Add morelike-based topic matching functionality to the backend
Add a Topic class for representing topics, and integrate it with the
ConfigurationLoader and TaskSuggester, and expose topics to JS as a
a ResourceLoader package file (homepage/suggestededits/Topics.json).

Task configuration is loaded from $wgGENewcomerTasksTopicConfigTitle,
which works much like the task type config; a working example is at
[[mw:Growth/Personalized first day/Newcomer tasks/Prototype/topics/cs.json]]
For now, human-readable task names are also loaded from that file;
that can be revisited once we have the list of ORES-based topics.

To keep PageConfigurationLoader simple, the logic related to page
loading is moved to a new PageLoader class, so ConfigurationLoader
can focus on config validation and high-level loading.

It seems that using morelikethis: overrides srsort=random and the
returned result set is deterministic and would be the same for all
users. Dealing with that is left for a follow-up patch.

The patch also includes some small improvements to configuration
loading in general:
- use in-process caching on top of the existing cluster-local caching
  (successful results are already in-process-cached via properties in
  PageConfigurationLoader so this will only affect loading errors)
- simplify dependencies by using RevisionLookup instead of WikiPage
- simplify JSON parsing code

Bug: T240512
Change-Id: I2bf6198f6f4a57ba3bd82bc0dfdd99dde70ffe78