Fix task card CSS height calculation
The task card height is set in em units, so it can adjust to
the font size being different on some wikis. On mobile the
task title font size is 1em and the text font size is 0.8em
but for desktop the title font size is 1.2em and the font size
is 0.875em so we cannot use the same height.

Since height is used several times through the CSS code, rather
than try to override everything for mobile and make the CSS even
harder to read, the patch introduces a mechanism to dynamically
declare LESS variables, and declares the height based on the skin.

Bug: T238280
Change-Id: I18e92d1123177e5a08f00c81c125a5bdaacde4b8