Suggested Edits: Add topic filter button
This patch begins to implement the spec for the topic filter button. In
production, the patch is a no-op because it depends on
GEHomepageSuggestedEditsEnableTopics being set to true.

If the feature flag is set to true, a second button widget is prepended to the
filter button widget group.

Follow-up patches are needed for:

* Re-using updateButtonLabelAndIcon method for setting the icon/label for
  topic button, once the dialog for setting topics exists
* Set the pulsing blue dot if the preference has never been set
* Opening a dialog when the button is pressed (T238612)

Bug: T238611
Change-Id: Iaee62ee7cb81854177f84a893aaedcfc408f67b8