Suggested Edits: Add pulsing blue dot if user never set topics
The pulsing dot should show only if the user never set topics. If the user picks
topics, saves their preferences, then unsets topics later, the pulsing dot
shouldn't show up and the progressive flag should not be set.

This is done by checking for an empty string from the preference (never set)
or an empty array (was set but is now empty array).

For discussion about placement of the dot on desktop, see T238611#5789718 and
T238611#5790124

Bug: T238611
Change-Id: I7484350199744cc686ccb94dbe06e3cbe64cb27b