Newcomer tasks: Make a separate search query for every topic
As a temporary solution to issues with query length limits and
weighting, use a separate query for every topic. Also use the
chance to set the topics of suggestions, which was previously
not possible.

Since this is a temporary workaround, no effort is made to deal
with duplicates resulting from the same article matching multiple
topics.

Bug: T242560
Bug: T242476
Change-Id: I27be56b0380fa37d8d64363dc9a19e33b7dd9409