Handle and log task topics and scores
Adds the concept of topic scores to the Task object, sets the
scores in SearchTaskSuggester, exposes them and topic IDs in the
API and adds them to the card log data.

With local search, the score is just the overall search score
(we are taking advantage of making separate searches per topic;
otherwise we couldn't get per-topic scores). There is no way to
get scores via the API, so for remote search we just put in a
fake score to aid testing.

Bug: T242052
Change-Id: I75dcf39e8d340682cb78049d64d11d6eb66318c9