Suggested Edits: Perform state updates even when presets is not an array
Instead of checking whether this.config.presets was an array, this was
checking whether it was empty. That caused buggy behavior when the user
manually unselected all topics: the dialog state would not be cleared on
cancel if the initial state was empty (no filters selected).

To fix this buggy behavior, we always need to do the state update, even
if this.config.presets is not an array. Instead of returning early,
treat it as an empty array.

Follow-up to 87ccf0171dec356e13d7489db0a2920678877de4

Bug: T242818
Bug: T238611
Change-Id: I8f0a23b54bd6c811046915c22addca8896ebca5e