Make sure to show mobile homepage discovery notice at least once
The notice is always shown when the welcome survey is in the
'source' URL parameter or in the referrer. Add logic to also
show it on any other pageview, if it has not been shown before.

The main effect of this will be to show the homepage notice to
users who have registered on desktop once they first use mobile.
It will also make sure the notice shows up to users who start
on mobile if they leave the welcome survey by directly editing
the url, or just abandon the browser tab alltogether and later
return to the site in a different way.

Also fix a minor error where users got the wrong notice if they
abandoned the survey by going to their homepage.

Bug: T224883
Change-Id: Icd64ebff45f12c7b4bb188d9333d1f737ad0f948