Do not log a close event on suggested edits initiation
Only log when navigating to the overview page, not when jumping from
one module to another.

The start-editing -> suggested-edits transition uses replaceState so
that navigating back brings the user back to the overview, so this
won't prevent the user backbuttoning out of suggested edits to be
captured as a close.

Bug: T242052
Change-Id: If58a46fe5cc76930034e624466557ca34aa80118