Do not call render methods when the module might not be set up
Split out shouldRender() from canRender(); use the latter to
signal whether rendering the module is a valid operation at
all, and use shouldRender() for what canRender() was used: to
tell whether it should be output in the homepage HTML.

This fixes the problem that getJsData() was always called,
even when the module might not have been functional
(e.g. Mentor module when the user did not have a mentor).

Also fix an unrelated problem that SuggestedEdits::canRender()
tried to load the configuration even when the module was not
enabled.

Bug: T243448
Change-Id: Ifcfbe8b0c1ac82f2642f0296c410b2028bf01a47