StartEditing: Handle configuration errors more gracefully
If topic matching is enabled (default config in the extension and WMF
production) but the topics page is not configured, then topicData returns an
error instead of data to be used for populating suggestion widgets. This patch
logs the error from configuration loader in the browser console and then
displays the start editing dialog as if topic matching were disabled.

Change-Id: Ic5e8e9c5e1d97486b26d5f1cb1b47053a68778c3