Homepage desktop and mobile style updates
This patch implements the updates for desktop and mobile styles of
Special:Homepage. Structural changes include:

* add mentor icon to mentor module
* change start account icon to user avatar
* for start module, don't use check marks for completed state on summary view
but do use check marks in detail view
* add language direction to overlay data
* move change email link on to single line

Then there are many style updates mostly following the proposed changes in the
phab tasks but often deviating from them to fit the LESS styleguide
specifications and/or for accessibility.

For follow-up (or fixing in this ptach):

* Adjust dark blue highlight box shadow on mobile to extend to width of viewport
* the count of edits in the start module mobile detail view is not aligned to
  the left

Bug: T232546
Bug: T239442
Change-Id: I595abc5cb05ac1fa611a551adab911b7fb535000