Use AQS data from two months ago
AQS processing happens monthly, and is a manual process, so at the beginning of
the month (i.e. today, Feb 4) there is no data available for January 3-4. To be
on the safe side, let's use data from two months ago.

For a follow-up, we can update the suggested edits module to have some different
UX when the number of edits returned is 0.

Bug: T244164
Change-Id: I3d0362bdfc1fa19255ff385203704e5d606d7b2a