Remove state from PageLoader
Instead of giving it a title during construction, pass it in the
load() call and reuse the same PageLoader for all config pages.

This was always the saner approach, not sure why I didn't do it like
this originally. It will also make it a bit simpler to add separate
configuration pages for morelike and ORES topics, and to validate
configuration page edits.

Bug: T243477
Change-Id: I0a6d2ab1aa1ac4ca167714379cfe646049ac6192