Sort ORES topics alphabetically
Topics in an earlier group always precede topics in a later group
(the idea being that on the client side you can reconstruct the
group structure by just iterating through topics and pushing new
groups in an array as they show up). Topics in the same group are
sorted via Collation.

Unlike category sorting, there is no way to provide a sortkey
manually; it would be easy to add (other than having to move the
configuration out of mw.org since it would become wiki-specific)
so we'll deal with that when it turns out to be necessary.

Bug: T246061
Change-Id: Iebaf52fd29101c70a590dc9ba691a232d0e08af0