Guidance: Add pulsing blue dot on desktop and mobile
This patch loads the Help Panel when clickId is present, and adds the guidance
JS/CSS to the help panel module. The JS and styles are for desktop and mobile,
as well as for VE single tab and separate tabs.

The blue dot should always show up after a user clicks on a card in the
suggested edits module, unless the user has already press "Edit" once for that
task type in that particular skin.

Bug: T244435
Change-Id: Ia5adc8d445b38e839c4bc29dd3f4662aa6ec6e8d