Help panel: Add root screen buttons
This patch implements the root screen version of the help panel which will be
used with providing guidance to users of suggested edits. The patch reworks the
help panel to use the tree structure but does not add any guidance-specific
code.

Animation for transitions between panels will be dealt with in T247999

The suggested edit button will be added in T248001

Bug: T244546
Change-Id: Iedc91fd275a1990ed75ed99f341fb8896e416242