TopicFiltersDialog: Actually prevent superfluous 'search' events
We were trying to prevent these events from being emitted, but it wasn't
working because the event handler was debounced, and checked
this.updating too late. Fix this by checking this.updating immediately,
and only calling the debounced handler if it's false.

This fixes strange bugs that were happening when closing the
TopicFiltersDialog while any card but the first card was being
displayed.

Bug: T247375
Change-Id: Ic52c45d22908aa7583c3de595f7d46a7417749a3