Help panel: Initial work on suggested edits panel
This patch adds some scaffolding for the suggested edits panel within the help
panel when guidance is enabled. Since this is behind a feature flag that is not
enabled in production, the code should be OK to merge as is.

The patch reworks the mobile peek contents to be reusable both in the mobile
peek and in the help panel process dialog. It also implements a panel layout
widget for the suggested edits screen that just loads the header for now.

None of the guidance specific behavior (e.g. when the help panel loads after
clicking on a suggested edit task, the panel should open automatically to the
guidance screen) is implemented yet.

A header exists for quick start tips but the content for that is complicated and
will be done in a subsequent. The footer will be added in a follow-up patch as
well.

Bug: T244546
Change-Id: I27d2cd2bc4ef545fccdd392dfcac943430854725