Suggested edits: Suppress VE dialogs for suggested edit page views
If we arrived on a page through suggested edits (detected using the
geclickid parameter), tell VE to disable the welcome dialog and the user
education popups (pulsating blue dots).

For the user education popups, we have to do this by pretending that the
user has them disabled in their preferences, but not actually saving this
preference change. For the welcome dialog, we can simply call
disableWelcomeDialog().

Bug: T235566
Change-Id: I94ab75a27a7c34c899224f7cb1e19b2a68aa7f1f