Add post-edit dialog widget
Adds a dialog to be shown after saving a suggested edit,
containing a new task suggestion. Actually displaying the
dialog is left for a separate patch; for now it can only be
triggered manually with

mw.loader.using( 'ext.growthExperiments.PostEdit' ).then( function ( require ) {
    var PostEdit = require( 'ext.growthExperiments.PostEdit' );
    PostEdit.setupPanel();
} );

Instrumentation is also done in a follow-up patch.

Bug: T245790
Change-Id: Ie15d376315301ecd9c177ff453cd799a5c01040a