Help panel: Add quick start tips to guidance
Features/functionality:

* Add REST handler for getting tips
  * TipsBuilder service used with the REST handler
  * rudimentary api-testing added for local runs, it's not ready for enabling in
    CI as we don't have (yet) the ability to override the ServiceWiring in CI,
	see also discussion in Ie155223f8d33ce9cc84d6fdcac8941e4f22f1550
* Front-end code calls the API to load tips based on current editor
interface and task type. In this patch there is no attempt made to deal with
errors from the API or implement a loading indicator if the API is slow; those
can be done as follow-ups
* "Tips" are modeled as a collection of "Tip Sets" (each number in the UI
is a Tip Set), and each Tip Set is composed of individual Tips. Each Tip has its
own configuration and renderer
* The tips are constructed using a structured pattern for message keys, that
follows the pattern of prefix + skin + editor + task type + tip type + step
* When the tips are rendered, there is a mapping that associates the message
keys with additional parameters needed to render those messages. The mapping is
a set of closures that return the parameters needed for rendering.
* Tips are rendered with fallbacks for the messages. VisualEditor and Vector are
treated as the default fallbacks. So if the request is made to render a
particular message for copyedit tipset '2' for 'minerva' and 'wikitext', we fall
back to checking 'minerva' and 'visualeditor', and if that message doesn't exist,
we fall back to 'vector' and 'visualeditor'. If the default message doesn't exist,
the tip is not rendered.

Follow-up tasks:

* Advancing the IndexLayout via a timer (T253009)
* Instrumentation (T253010)
* Handle failures with API query (T253011)
* Get the API tests running in CI (T253015)

Bug: T244541
Bug: T245786
Change-Id: Idce0b0736505d939759a33beda561777fab1343e