Make Mentorship::getMentor return either bool or User
According to the previous code, if $this->mentor was
null (the default state), and Mentor::newFromMentee
returned false, Mentorship::getMentor will return null
althrough the docs don't allow for that.

This is an unlikely case and shouldn't happen,
but I think the code should account for unlikely
things as well, and never let a method return something else
than the docs allows it to.

Change-Id: I6075ca6bb3e916235a70cc94859b35314a1e1acf