Help panel: Animate transitions between panels
Move the home panel out of the StackLayout, and position it to the left
of it. Initially, the home panel is visible and the StackLayout with the
subpanels is out of view to its right. When the user navigates to a
subpanel, both the home panel and the StackLayout slide to the left,
until the home panel is out of view. When navigating back to the home
panel, the reverse happens.

When navigating to/from the suggested-edits panel, the blue background
in the header animates similarly. The contents of the header itself do
not animate, they change immediately and in-place.

Move the footer of the suggested-edits panel from the $foot to the
bottom of the $body. This allows us to keep the body height constant.
The footer doesn't animate in with the rest of the panel contents, but
that's OK: the footer will be hidden in edit mode anyway.

Also update the body height calculation: now that the home panel is
separate, it's always measureable, so we don't have to look at whether
it's the current panel and only measure it if so. This will also allow
us to open the dialog to a different panel.

Bug: T247999
Change-Id: I52b121c707a962c3caf7c12ca4fc3e032a61b3cd