Resize image card thumbnails to be exactly 260 pixels
When the API returns a thumbnail that's narrower than the
task card, but the original image is wide enough (this can
happen with RESTBase where we have little control over
thumbnail size), alter the thumbnail URL to cover the card.

Thumbnails wider than the card are left as-is (and will be
resized by the browser). Thumbnails where the original itself
is narrower than the card (this is common for logos) are
still displayed with background-repeat. No attempt is made
to make the thumbnail as wide as possible (MediaWiki's image
URL handling would make this more complicated than it's worth).

Bug: T244210
Depends-On: I77c68c99a708d3e0a91de31d8a95e217230d2734
Change-Id: I2b0d7e2a87404788810cd310344a9d650cfc8b48