Suggested edits: Make header+tips scrollable together
Only the tips panel was scrollable, but we want the scrollable area to
be the header panel plus the tips panel. To achieve this, wrap the
header and tips panels in a StackLayout.

This also makes the CSS simpler: instead of using flexbox, we can now
just use height: calc(100% - 40px);, since the height of the footer is
fixed at 40px.

Bug: T253678
Change-Id: I9e1b82ab3c698109bc8d248768d13b44935e6ac0