Improve suggested edits card image positioning
Instead of showing the image from the top, center its top half
(ie. show it with background-position-y:25%). This works much
better for faces and it's not any worse for most other images.

Also refactor slightly to handle more CSS properties non-dynamically,
and wrap the thumbnail URL in quotes (this shouldn't matter as
our thumbnail URLs are properly URL-encoded, but it prevents loading
failures if one isn't).

Bug: T244210
Change-Id: I4e83db8ba3832f946f7fd7b8819c15dc56662c83