Help panel: Update guidance behavior rules
Business rules:

* In guidance, track the help panel state so we know which panel the user had
opened and which tip was selected. Persist this across page reloads, tab
switches, etc.
* Also, on page reload, open the help panel if it was open before or keep it
minimized if it was minimized
* When guidance is available, ensure that the help panel button is always
available, even in Read mode (unlike traditional help panel)
* When toggling between Read/Edit/Edit Source, show/hide the footer that says
"Ready? Click/tap editing..."
* If the user in read mode changes the state of panel, by toggling tips or
navigating which screen they're on, then switching modes should not change the
state of the panel. If they have not touched the help panel in read mode then
we would want it to animate back to the root screen upon switching to edit mode.
* Set locked mode for the help panel if the current panel is suggested edits and
the page is in read mode.

Code changes:

* In SuggestedEditSession, keep track of 1) if the help panel should be open, 2)
the current panel shown, and 3) if it should be locked.

* Remove the pageMode concept and suggested-edits sub modes (read/edit), instead
there is a method called updateEditMode which callers can use to tell the help
panel that the page is being edited (not read), and that relies on the help
panel loggers' isEditing() method. updateEditMode() also handles switching the
panel depending on whether the user has interacted with the guidance screen.

* Show help panel in read mode if there is an active session. The help panel
normally does not show in read mode on an article. But if the user has an active
suggested edit session, let's allow it to display. Also ensure that mobile peek
is only shown once with this change.

* Store current tip tab in state. Arguably we should not be doing this via the
call to getTips() but given that this call is already generating the
StackLayout rather than only retrieving data, setting the tab panel makes sense
to do there as well.

* Rework QuestionPoster.js to rely on SuggestedEditSession as it makes the rest
of the code cleaner, even if conceptually it's a bit odd. The module is already
loaded so there is no performance penalty in using it.

Misc:

* wikipage.editform passes an object, not a string, and we need to pass a string
to isValidEditor

Bug: T244431
Change-Id: I39fa8fc84b4be6441333a740a62fbd56c34d75df