Help panel: Adjust primary actions / settings cog logic
A previous commit (77fb9e9) switch the buildSettingsCog method to empty the
primary actions and then rebuild on panel switch. That broke the current
existing use of primary action when a question is posted and again when the
question posting is completed.

Longer term, it would make sense to move the settings cog out of the primary
actions so we don't have to hack around this. Alternatively we could build the
settings cog once and toggle the particular items (this will be easier when
it's converted into a proper OOUI menu). There might also be business rules
updates for when particular links in the settings cog are shown.

Given all the above, this less-than-elegant patch seems like the best way to
solve this problem right now.

Bug: T254410
Change-Id: Idf0b6b51e0bf717da36130897807a6038a7d3364