Remove homepage pageview token verification
The homepage pageview token had a cryptographic signature which was
too clever for its own good. The scenario it tried to avert (an URL
with a click id getting somehow shared or reused, messing up logging)
is pretty unlikely, and it prevented us from generating an URL with
a click id on the client side, which we need for the post-edit panel.

Bug: T254416
Change-Id: Ib08e0e388c782c1612d458ffd55a7e895bda2f42