Change "language" to "text" in the help panel
In English, the word "language" may have the senses of
"writing style" or "specific words". The same is not necessarily
true for other languages: In Russian, Ukrainian, Hebrew, and Portuguese
the word for "language" in the sense of "systematic means of communicating
ideas or feelings by the use of conventionalized signs, sounds, gestures, or marks"
does not also have the sense of "style" or "specific words".
(The same is probably true for many other languages, but these are
the ones I checked.) Even in English this sense may be confusing,
especially on a very multilingual site such as Wikipedia.

It was already translated to several languages as "language",
in the sense of "foreign language" or "native language", and it
was done incorrectly.

It is therefore better to use a less ambiguous word in the message.
I recommend "text", but it's also possible to say "phrasing", "style",
"prose", "words", or something else.

Change-Id: I445f13bfe3762bad37dc154cefc0fbd437d6f773