Help panel: rearrange close icon and cog menu
Remove the cog menu and replace it with a close icon, and a plain list
of the menu items on the home panel. This also lets us remove all special
handling of actions.

To get the close icon to the right place without manual hacks, we need to
make it the primary action. That is a bit weird, but has no side effects
other than making close the default action, which is fired when pressing
Ctrl-Enter, which is not really a meaningful key combination on any of the
panels other than the question poster (which does not have a close icon).

Bug: T255254
Change-Id: I7c4fe8cf56bb858e2fe13b19ddcb61828c946404