Help panel: Add warning if user is in non-preferred editor
* Add SwitchEditor panel to render in the suggested edit screen
* A new sitewide configuration value is introduced to define the preferred
editor for suggested edits, with a default value to "visualeditor"
* The link to toggle between editors uses VisualEditor's library to initiate the
switch
* Mobile switching is not implemented. Users are opted into VE by default
so they should understand how to switch back when the warning notice
instructs them to. Providing a link to initiate mobile switching would
require some changes to MobileFrontend, and that's left for a follow up.
* Instrument the link which initiates editor switching; the rest of the event
logging code will provide additional context (including the current editor)

Bug: T255074
Change-Id: Iaf349773d1e71731ce50b39c7ecfca51d161de38