Help panel logger: Set the editor interface via hooks
* When MobileFrontend fires hooks for opening and closing the editor, set the
editor interface (visualeditor or wikitext) in the help panel logger
* Update isEditing() to check if this.editor is non null, and modify getEditor()
to use this.editor if it's already set; this works around a problem we have
where getCurrentEditor() attempts to find the text area in the DOM but we
detached it in order to show the help panel.

In a follow-up we could probably handle desktop editor switches via the 
hooks used for editor opening/closing.

Note that two 'impression' events are logged when reloading a page with guidance
enabled, but this is also the case with the main branch currently in 
production. We should fix that in a follow-up.

Change-Id: Icb7d5d44f6da900dee0d6fa7955d50270706738d