Help panel: Make the CTA button a toggle
Instead of hiding the button while the help panel is open, change its
icon from a question mark to an up chevron, and make it close the help
panel when clicked. To make room for the button, position the help panel
farther up.

Also remove the text from the button and make it round, and increase its
font size: it was 12.25px instead of 14px, because there were two
font-size: 0.875em; rules layered on top of each other.

Bug: T255255
Change-Id: Ia1fc8b3d649568fca91ace70e3d88139c9444d4c