Newcomer tasks: UI fixes to the text elements on the post-edit dialog
Alongside some smaller style changes, redo the flex model for the card
text. Old model was horizontal flex with line wrapping, with the goal
of allowing pageviews and task type to take up the full line when they
are long and there's space (no description). Since that's not wanted,
and  for some reason right-justifying the task type refuses to work
when there's no pageview data, use a more straightforward layout with
a vertical flex for title, description and a container which holds a
horizontal flex for the pageviews and task type.

Bug: T255825
Change-Id: I0b715ee6952885f89b66620b61dd10fd7c48d6b0